blog_project/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-ckeditor
